refactor(data): 👷 metadata in correct folder
