+++ app/work-experience/komercni-banka/page.tsx
@@ -3,9 +3,9 @@ import { Metadata } from 'next'
 import ProjectPageLayoutWrapper from '@/components/layout/project-page/ProjectPageLayoutWrapper'
 import PageNavigation from '@/components/page-navigation/PageNavigation'
 
+import { metaDataKomercniBanka } from '@/lib/data/metadata/projects/work/metaDataKomercniBanka'
 import { sections } from '@/lib/data/projects/work/komercniBanka'
 import { projectsWorkReact } from '@/lib/data/projects/work/projects-overview/react/workReact'
-import { metaDataKomercniBanka } from '@/lib/utils/metadata/projects/work/metaDataKomercniBanka'
 
 import { TEXT } from '@/localization/english'
