+++ lib/data/metadata/metaDataShared.ts
