+++ lib/data/metadata/projects/personal/metaDataCryptomania.ts
