+++ lib/data/metadata/projects/work/metaDataKomercniBanka.ts
